move DOM manager into own module
